// ui/src/api/models/IMediaUrl.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type IMediaUrl = {
    type?: IMediaUrl.type;
    displayName?: string;
    url: string;
};

export namespace IMediaUrl {

    export enum type {
        IMAGE = 'image',
        VIDEO = 'video',
    }


}